Updated protobufs with hack/update-proto.sh
This patch updates the protobuf files by running hack/update-proto.sh.
It appears this was not executed as recently as necessary given the
number of changes based on no changes to the source code. That or the
version of protoc is sufficiently different on this system (it was
installed with "brew install protobuf").