Update Listener's Port field to list
This patch updates "Listener.Port *int32" to "Listener.Ports []int32".
This change is based on the existing documentation for the field, which
clearly indicates the field is meant to be a list. Additionally, it
makes sense for the field to be a list as a single listener should be
able to listen on multiple ports.