Adding CRD default for GatewayClass status
Accomplishing this required upgrading to the newer v1 API version for
CRDs. This does mean that these CRDs would only be able to be installed
on clusters with Kubernetes 1.16 or newer. This also updates the example
to be valid with our updated spec, and renames the CRD files to match
our updated x-k8s.io API version.