add recommendation to add finalizer to GatewayClass
There is nothing that prevents a user to delete a GatewayClass resource
when there are Gateway resources that refer to it.

This patch adds in recommendation to use finalizers to prevent deletion
of GatewayClass resources that are in use.

This is a secondary safety-net. The first one is based on our personas
and RBAC where for a large cluster, only a small subset of users should
have delete permissions on GatewayClass.

See #170