Make an editorial pass over the API documentation.
Especially now that enum constants are rendered in the HTML API
documentation, the godoc conventions make it harder for a reader
to understand the intentions of the documentation.

To handle passing a literal "*" character through the two layers of
Markdown, we postprocess the`"*"` pattern into a HTML entity.

Make an editorial pass that improves readability without any semantic
changes.

Signed-off-by: James Peach <jpeach@vmware.com>