Fixing API Reference Doc Generation
Previously reference docs were converted from godoc to HTML with
gen-crd-api-reference-docs and then converted to markdown with mkdocs as
part of generating the docs site. This takes out the second step and
instead inserts the generated html for the reference docs directly into
the generated html for the docs site.