Drop minLength=1 from Group
When using core API, the apiGroup is "" (empty strings).

For example:
```
$ kubectl api-resources
NAME                              SHORTNAMES   APIGROUP                       NAMESPACED   KIND
secrets                                                                       true         Secret
```

However, current CRD does not allow empty strings and it seems to
expect "core" is specified as per example.
It is different from other Kubernetes resources, such as Role, so
difficult to find it.