Fix v1alpha2 "HTTP routing" example
The "bar-route" HTTPRoute had a bug where both backendRefs belonged to
the same rule so traffic will either match the header in which case it
will be split 50/50 between bar-svc-canary and bar-svc, or it won't
match the header in which case I'm not sure what happens.

This PR adds a second HTTPRouteRule to the bar-route HTTPRoute so the
yaml manifest matches the descriptive text.

https://kubernetes.slack.com/archives/CR0H13KGA/p1642527563035400