Allow the user to tell the conformance tests leave config in place
This is a trivial change because almost all of the work had been
done. Only the last step wasn't implemented.

To run the conformance tests and leave the configuration in place at
the end of the run:

 $ go test ./conformance/conformance_test.go --gateway-class={your-gw-class} --cleanup=false

This is helpful when a test fails because you can then explore what
went wrong.