mkdocs: enable "line-based comments"
We would like our example YAML files to include comments that do not
show up in the guides where the files are used as jinja templates,
while still ensuring they are valid YAML files.

Configuring the jinja Environment with "line_comment_prefix" allows
you to control the syntax of the jinja comment directive. If we set
that to "##", then such lines are ommitted during jinja templating,
and are treated as YAML comments.

See also
https://jinja.palletsprojects.com/en/2.11.x/templates/#line-statements