Ensure that "Supports:" statements stand alone
Some of the "Supports:" statements in the doc didn't have an empty
line preceding them so the markdown engine squashed them into the end
of the previous paragraph. More diligent people than I probably
noticed this but I missed one completely so this change ensures that
each is its own paragraph. In most cases this is simply adding a blank
line.