Update examples to use v1beta1
Where v1beta1 resources exist, this updates the documentation to use
these instead.

The exceptions to this are:

- The "introducing v1alpha2" blog post still uses v1alpha2 examples for
all examples.
- All GEPs are unmodified
- Any examples that need to reference _some_ v1alpha2 resources, will
continue to use v1alpha2 resources for all (even where v1beta1 versions
exist). For example, the example in guides/tcp, the Gateway in its
example is a v1alpha2 resource, as TCPRoute does not exist in v1beta1.
This matches the general advice in [slack][api-advice] to continue to
using the alpha API for these cases.

[api-advice]: https://kubernetes.slack.com/archives/CR0H13KGA/p1656533075102939