Update Spec Links to link to v1beta1 resources
Previously, all links to the API specification pointed to the v1alpha2
versions of each resource. This updates the documentation to point to
v1beta1 instead.

NB: guides/tcp has been explictly ignored; it will continue to point to
the v1alpha2 variants, even for things like `Gateway` and `HTTPRotue`,
which have graduted to v1beta1. This is because the general
[advice][api-advice] is to continue to work with v1alpha2 versions if
you want to support alpha-features.

[api-advice]: https://kubernetes.slack.com/archives/CR0H13KGA/p1656533075102939