conformance: Update HTTPRouteInvalidParentRefNotMatchingListenerPort
The manifest for not matching listener port is rejected by admission
webhook due to missing port for Service reference. While the test is to
verify some other behavior (i.e. parent gateway port is not matching),
it's better to have valid resource to avoid any confusion or any
implementation expectation, not to mention to be consistent with other
conformance tests (e.g. invalid-cross-namespace-parent-ref).

This commit also updates gateway name used in verification step to be
consistent of the gateway used in HTTPRoute resource.

Signed-off-by: Tam Mach <sayboras@yahoo.com>