gateway.listeners: make optional
This changes `listeners` field to be optional (i.e. allow 0 length
list).

The motivation here is that a Gateway actually provisions underlying
infrastructure. This can take a long time, too, depending on the set.
 It may be useful to initially provision a Gateway with
*zero* listeners, and then later add them.

This PR loosens validation to allow not configuring any listeners, with
the expectation that underlying infra would be provisioned, then a user
would add specific listeners.