chore: remove references to conformance profiles
There were a few references to conformance profiles in the
godoc for the conformance test suite features. These references
existed because this work originally came from the conformance
profiles project branch which was work in progress, however since
this change was cherry-picked out it needs to stand alone so
referring to the not-yet-added conformance profiles work would be
confusing.