validation: take parentRef.port into account
Currently, we cannot have:

```yaml
  parentRefs:
  - name: gw1
    port: 80
  - name: gw1
    port: 8080
```

This seems like an oversight; our validation only takes sectionName into consideration.

Note: this DOES allow overlapping parentSection selection. For exampe, i may select port=80 and sectionName=http
which refer to the same thing. This is not something we can catch in
validation. Perhaps we need to clarify the behavior here in the spec, if
its not clear. Note this could always happen with 2 different routes
anyways, though.