conformance/utils/kubernetes: Return more errors from DumpEchoLogs
Errors from streaming pod logs or reading logs were invisible to users
running the suite since on error the loop through pods just continued.

This test has been flaky in CI so it would be good to know if that is
caused by errors in the request fetching logs or otherwise.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>