conformance: Ignore state of terminating pods
In NamespacesMustBeReady helper that checks if all pods in a namespace
are ready, used to gate test runs to ensure we're not making requests to
backends that are still coming up, etc.

As tests create/delete Gateways (and associated infrastructure
components, e.g. control/data planes) in commonly used namespaces, there
are often Terminating pods that the helper waits for, which
significantly increases test runtime and causes timeouts.

Instead, if a pod as a deletion timestamp, indicating it is being shut
down, we ignore that when checking the pods in a namespace.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>