conformance-profile: Remove experimental build tag
This is to make it easier for some implementation that imports the
conformance test framework and run tests in the same codebase.

Understand that cleaning up of similar/duplicated code is required to
have a clean conformance test framework, however, such work might delay
v1.0.0 GA release, hence, I only remove the build tag to remove the soft
blocker in this PR. Subsequent cleanup/optimization can be done later.

Relates: #2242
Signed-off-by: Tam Mach <sayboras@yahoo.com>