Fetch resources in bulk instead of single API calls as the default.
Eg. When we have a set of HTTPRoutes and we want to fetch all Backends
referenced by this HTTPRoute, it's slightly better to fetch all the Backends
with one API call and then filter these out based on which are being referenced.
This is in contrast with fetching each individual backend through a separate API
call that is referenced by the HTTPRoutes. (Even at smaller scales, multiple API
calls starts to seem inefficient)