Lazily fetch events instead of making them part of the ResourceModel.
Events associated with a resource need to be fetch independently for each
resource. This means the number of API calls grows linearly with the number of
resources that we have.
- Pre-fetching events for all resources BEFORE PRINTING ANYTHING gives the
  impression of the CLI being unresponsive.
- Lazily fetching events allows us to print the describe-view for a single
  resource fully before fetching events for the next one, thereby giving the
  user an impression of the CLI making progress.

Also, it makes sense for Events to not be part of the ResourceModel given events
are just tied with a single resource and have no interdependence (so don't need
some pre-calculations)