gwctl performance improvements (#3145)
* Fetch resources in bulk instead of single API calls as the default.

Eg. When we have a set of HTTPRoutes and we want to fetch all Backends
referenced by this HTTPRoute, it's slightly better to fetch all the Backends
with one API call and then filter these out based on which are being referenced.
This is in contrast with fetching each individual backend through a separate API
call that is referenced by the HTTPRoutes. (Even at smaller scales, multiple API
calls starts to seem inefficient)

* Lazily fetch events instead of making them part of the ResourceModel.

Events associated with a resource need to be fetch independently for each
resource. This means the number of API calls grows linearly with the number of
resources that we have.
- Pre-fetching events for all resources BEFORE PRINTING ANYTHING gives the
  impression of the CLI being unresponsive.
- Lazily fetching events allows us to print the describe-view for a single
  resource fully before fetching events for the next one, thereby giving the
  user an impression of the CLI making progress.

Also, it makes sense for Events to not be part of the ResourceModel given events
are just tied with a single resource and have no interdependence (so don't need
some pre-calculations)