HTTPRoute: allow more match clauses
Today, a route is limited to 16 routes with 8 matches each. This is
problematic in real world environments.

While its easy to split a route up (you can easily get the same
behavior, and similar cognitive complexity by having 1 route per
HTTPRoute), its quite hard to split up matches. For instance,
https://github.com/istio/istio/blob/d7a9700d5eaa4e9728274b408623670f48deadb5/samples/bookinfo/gateway-api/bookinfo-gateway.yaml#L23-L38
is an example of a route matching the exposed APIs for a frontend. This
is only a small tivial example, and it already uses 5. In our user
base, we often see far beyond 8.

Splitting these up is both complex for the user, and may actually lead
to different behaviors if implementations treat route groups differently
(for instance, if we add retry budget -- that could now be split; other
core or implementation specific policies may behave similarly, this is
just an example).

The limit on 8 here seems quite small given the context of real usage,
the other limits in the API (16 routes, 64 listeners), and the cost of
manually working around it. Based on this, I propose we raise the limit
to match Gateway listeners.