conformance: Update the Gloo Gateway project's published 1.0.0 reports (#3289)
* Add the Gloo Gateway v1.17.4 conformance report

This commits updates the published 1.0.0 conformance suite
reports for the Gloo Gateway project.

Signed-off-by: timflannagan <timflannagan@gmail.com>

* Remove the stale v2.0.0-beta1 Gloo Gateway conformance report

This commit removes the published v2.0.0-beta1 Gloo Gateway
conformance report. Previously, support for the Kubernetes
Gateway API integration in the solo-io/gloo repository was
done in a v2 feature branch and this tagging convention is
no longer relevant.

Signed-off-by: timflannagan <timflannagan@gmail.com>

* Update the Gloo Gateway project implementation status

Signed-off-by: timflannagan <timflannagan@gmail.com>

---------

Signed-off-by: timflannagan <timflannagan@gmail.com>