docs: Use relative links to site pages
This commit modifies the documentation to use relative links between
pages on the site instead of absolute URLs.

Linking to a page's relative file path allows MKDocs to ensure that page
exists and/or linked heading actually exists when building the site.

The URLs in the generated HTML files will be relative or absolute,
depending on the location of the linked file relative to a given page.