docs: Fix warning for invalid 'trademark' config
PR #3132 (commit 8285041) introduced a trademark disclaimer to the
website's footer, storing the trademark URL in MkDocs config under the
key `config.trademark`.

This is not a valid MkDocs configuration parameter and causes the
following warning to be logged when building the site.

  WARNING -  Config value 'trademark': Unrecognised configuration name:
  trademark

The correct place to store custom configuration is under the `extra`
key [1] in MkDoc's configuration.

This commit relocates `trademark` parameter to resolve the
configuration warning.

[1]: https://www.mkdocs.org/user-guide/configuration/#extra