docs: Fix several MkDocs info and warning messages (#3663)
* docs: Fix nav error in mkdocs.yml

This commit fixes the informational message logged by MkDocs due to an
incorrect format in the nav parameter.

> Config value 'nav': Expected nav to be a list, got dict with keys...

* docs: Update pymdownx.emoji extensions

This commit resolves the following informational from MkDocs that was
being logged due to the previous extensions no longer being supported.

  Material emoji logic has been officially moved into mkdocs-material
  version 9.4. Please use Material's 'material.extensions.emoji.twemoji'
  instead of 'materialx.emoji.twemoji' in your 'mkdocs.yml' file.

* docs: Fix warning for invalid 'trademark' config

PR #3132 (commit 8285041) introduced a trademark disclaimer to the
website's footer, storing the trademark URL in MkDocs config under the
key `config.trademark`.

This is not a valid MkDocs configuration parameter and causes the
following warning to be logged when building the site.

  WARNING -  Config value 'trademark': Unrecognised configuration name:
  trademark

The correct place to store custom configuration is under the `extra`
key [1] in MkDoc's configuration.

This commit relocates `trademark` parameter to resolve the
configuration warning.

[1]: https://www.mkdocs.org/user-guide/configuration/#extra

* docs: Ignore certain files from nav config

MkDocs currently logs an informational message stating that files under
the blog/ and gep-696/ (GEP template) paths are not present in the
site's navigation.

  The following pages exist in the docs directory, but are not included
  in the "nav" configuration:

This commit adds MkDocs's `not_in_nav` configuration parameter to
specify that these paths are intentionally excluded from the site's
navigation.