conformance-tests: fix flake in ExpectMirroredRequest (#3708)
Currently the Conformance Tests for request mirroring is flaky
because the mirrored request can sometimes not be found in the
log of the Pod.

```
Test:       	TestConformance/HTTPRouteRequestMultipleMirrors/1_request_to_'/multi-mirror-and-modify-request-headers'_with_headers_should_go_to_infra-backend-v1
Messages:   	Couldn't find mirrored request in "gateway-conformance-infra/infra-backend-v3" logs
```

The issue is that the logic that fetches the Pod log (`ExpectMirroredRequest`)
uses the current time (`time.Now()`) on every assertion-attempt as the
`sinceTime` parameter.

As a consequence, the assertion can miss messages from the Pods that have been
logged between the assertion attempts.

This commit fixes the issue by using the start time of the "check"
when fetching the logs (the same way it's already handled in
`testMirroredRequestsDistribution`).