Implement automatic NAV generation
This change makes the navigation section of GW API website
to be generated.

The modification adds a new Go program that is able to transverse
the GEP directory and generate a navigation section correctly from
the existing GEPs.

Additionally, some scripts and Makefiles are added to verify if the
generated nav.yml file reflects the current state of GEPs, a Github
Action that fails in case nav.yml is outdated