Automate GEP TOC generation and validate (#4075)
* Fix GEP API and geps metadata

The GEP API is using a wrong structure/tag name for some
fields, so this PR fixes the markers.

Additionally the metadata.yaml files from some GEPs are
fixed to reflect the right GEP API structure

* Implement automatic NAV generation

This change makes the navigation section of GW API website
to be generated.

The modification adds a new Go program that is able to transverse
the GEP directory and generate a navigation section correctly from
the existing GEPs.

Additionally, some scripts and Makefiles are added to verify if the
generated nav.yml file reflects the current state of GEPs, a Github
Action that fails in case nav.yml is outdated

* Add the new generated nav file