Removed invalid printcolumn marker from v1alpha2 ReferenceGrant
The v1alpha2 ReferenceGrant type is marked with +kubebuilder:skipversion,
meaning no CRD is generated for this version (it’s just an alias to v1beta1).
The +kubebuilder:printcolumn marker cannot be applied to a skipped version.

This issue was uncovered by the CRD marker validation added in the previous
commit. The printcolumn is already correctly defined in v1beta1 and does not
need to be duplicated here.

Signed-off-by: Killian Golds <kgolds@redhat.com>