Fix the API ref links (#4215)
Since we started to version the API reference, the links that
were previously pointing to the ref are broken. This happens because
there is no reference on the parent directory anymore.

To fix this, the development reference has been moved back to the
original location, so documentation can still keep working
pointing to the development version of api reference